Fix an issue preventing to use double values with scalar units.

+++ alien4cloud-core/src/main/java/alien4cloud/tosca/normative/Time.java
@@ -5,7 +5,7 @@ package alien4cloud.tosca.normative;
  */
 public class Time extends ScalarUnit<TimeUnit> {
 
-    public Time(long value, TimeUnit unit) {
+    public Time(double value, TimeUnit unit) {
         super(value, unit);
     }
 }
